itens do carrinho funcionando porém não deletando
por algum motivo está bugando e acho que tem a ver com a funcionalidade carrinho que antes apresentava o mesmo problema

+++ backend/src/main/java/br/senac/tads/petshop/dtos/ItemCarrinhoDTO.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -24,5 +26,5 @@ public class ItemCarrinhoDTO {
     private Integer unidades;
 
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
-    private Double subtotal;
+    private BigDecimal subtotal;
 }
